Fix .status.service naming collisions
Originating issue: [IBMPrivateCloud/roadmap#57621](https://github.ibm.com/IBMPrivateCloud/roadmap/issues/57621)

The first pass at implementing .status.service block for ODLM tried to
use the Kubernetes object name as a unique id for entries in the list of
managed resources. This assumption fails because names are only unique
among their GVK, not necessarily elsewhere. As a result, Service and
Deployment name collisions occurred, which led to new Deployment-related
state being written to the indices where Services of the same name were
stored, leaving the old Deployment state in the latter indices never to
be touched again.

This fix replaces all of the extra fuss around what the previous state
of the object was and instead replaces .status.service with the
.status.service that was most recently obtained from the current cluster
state and performs an update only if they do not deeply equal one
another.

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>