Remove dependency on Routes API
Originating issue: [IBMPrivateCloud/roadmap#58183](https://github.ibm.com/IBMPrivateCloud/roadmap/issues/58183)

Now, the Operator will only build with and handle Schemes related to
OpenShift if the Operator detects that it has been deployed on an
OpenShift cluster. Routes in this case specifically are skipped if
running on a non-OpenShift cluster.

Other Changes:

* Created function for getting the shared services namespace from a
  CommonService CR, which is a more straightforward approach than trying
  to list Authentications.
* Reorganized some utils-like files into common directories to clean up
  organization of pkg code.

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>