Fix registered API detection (#634)
Originating issue: [IBMPrivateCloud/roadmap#58413](https://github.ibm.com/ibmprivatecloud/roadmap/issues/58413)

Previously, whether or not to add the Route schema to the Manager was
determined by values set within the ibm-cpp-config ConfigMap that
indicated whether the cluster was OpenShift or CNCF. This approach was
vulnerable to timing issues that led to the Operator Pod repeatedly
crashing until the ConfigMap was made available.

Now, the code attempts to list Routes; if the error relates to the Routes
API not being registered on the cluster, then loading the scheme will be
skipped.

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>