Remove `state` parameter from platform-login Route (#638)
As part of authorization flow in OIDC, a state parameter is set to track
requests and defend against CSRF. When Common Services was using IBM
management ingress, it could define various nginx customizations within
annotations on its Ingress objects such as rewriting targets or even
scripted behaviors. Among other things, the IAM Operator's Ingress for
platform-login had been customized to generate a timestamp and submit
that as the `state` parameter for inbound authorize requests.

When Common Services moved away from using management ingress in favor
of using OpenShift Routes and the default OpenShift Router, it meant
that many of these customizations needed to be reevaluated and reworked
if they were to continue to be of use. In this case, generating
timestamps when requests arrived was not an option with the default
Router configuration, and so a hack was put in place so that the
authorize flow would continue to work in development.

This commit undoes that hack by no longer setting a state. The ID
provider will assign a timestamp and perform any additional redirects as
needed.

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>
Co-authored-by: Rashmi Khanna <k.rashmi.r@gmail.com>