Fix ownership of ibmcloud-* resources issues
Originating issue: [IBMPrivateCloud/roadmap#58841](https://github.ibm.com/IBMPrivateCloud/roadmap/issues/58841)

During backup and restore, the ibmcloud-cluster-info ConfigMap and the
ibmcloud-cluster-ca-cert Secret would block the IM Operator from coming
completely online because it expects that it owns those resources;
however, when the resources were restored, the ownerReferences were not
also restored. This fix to the IM Operator attempts to assert controller
status on behalf of the Authentication CR upon the ibmcloud-cluster-info
ConfigMap when it isn't controlled by that Authentication CR, and, in
the case of ibmcloud-cluster-ca-cert, the IM Operator will simply delete
the Secret and create a new one to replace it.

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>