Refactor Condition, Event, and Secret writes
There was a code comment in the reconcile logic about there being a
"strange issue" with writing Client status. Performing a write to the
Client CR on the cluster leads to overwriting the Client struct passed
to the update function with the resulting object on the cluster. This
means that changes to the status block in the Client CR are erased and
replaced with whatever the current cluster state is.

To address this, the Client CR status is determined and written after
any updates to the Client's other fields, which then avoids accidental
clobbering of the status when something else about the Client is
updated.

Additionally, the order in which the parts of a Client are reconciled
has also changed. Now, the Secret owned by the Client CR is written and
guaranteed first, and, if there is a later failure to register the OIDC
client or the Zen instance mapping, if applicable, the Secret is reused
for future registration attempts instead of creating a brand new Secret.
Client CR deletions were also updated to approximately mirror this
change, but in reverse.

Additional changes:

* Remove redundant ZenInstanceId check
* Remove redundant HTTP method constants already provided by net/http
  package

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>