Fix Dockerfile arch compile target (#732)
Originating issue: [IBMPrivateCloud/roadmap#61811](https://github.ibm.com/IBMPrivateCloud/roadmap/issues/61811)

The Makefile was using the wrong build argument to pass the `GOARCH`
value to the compiler for the manager. This resulted in the desired
value not being properly set for image builds, which then led to the
binaries only being compiled for Linux on whatever the host architecture
happened to be.

Given there are already separate Dockerfiles for each architecture the
project supports, the Dockerfiles no longer take this build argument
and, instead, the build stage in each file explicitly states the desired
`GOARCH` value.

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>