Return nil for deleting Client with empty ID
Previously, the `deleteClientRegistration` function returned an error in
the event that the Client CR passed to it did not have a Client ID set.
This led to Client CRs with no Client IDs and deletion timestamps
failing to ever be fully deleted/finalized.

Now, the `deleteClientRegistration` will return `nil` for the error value
in this case because an unset ID should mean that there are no
registered client credentials associated with it.

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>