Remove spec from ZenExtension type (#930)
Previously, a ZenExtension's `.spec` was represented as a
`map[string]string`. While the ZenExtensions that IM creates or looks up
do follow this structure, it is not technically enforced by the
ZenExtension controller. To avoid problems related to the lack of an
enforced structure, the `.spec` field no longer is in the IM Operator's
type representation of a ZenExtension, and it is instead read from an
Unstructured object.

Additionally:

* Operations on the `ibmcloud-cluster-info` ConfigMap that were
  previously happening during handling of the ZenExtension front door
  are now happening in the ConfigMap handling code.
* If the Zen host value cannot be obtained from the product-configmap,
  the cp-console defaults are used for `ibmcloud-cluster-info` fields.
* Remove redundant check for ZenExtension support

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>