Update MASTER_PATH based on existing SAML
If the `platform-auth-idp` ConfigMap does not have the `MASTER_PATH`
field, the reconciler will look for an existing SAML connection in the
database. If one exists, then the non-unified set of Routes is used for
completing the OAuth flows. If no SAML has been configured, then the
idauth unified Route is used instead.

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>