[FORD SAML LOOP SC2] - Update containers.go to add MASTER_PATH (#1009)
* Update containers.go

* Update configmap.go

* add idauth back for sc2

Signed-off-by: rashmi_kh <rashmi.khanna@in.ibm.com>

* introduce alt for provider

Signed-off-by: rashmi_kh <rashmi.khanna@in.ibm.com>

* Update configmap.go

* Update containers.go

* add master host to authSvc

Signed-off-by: rashmi_kh <rashmi.khanna@in.ibm.com>

* Update MASTER_PATH based on existing SAML

If the `platform-auth-idp` ConfigMap does not have the `MASTER_PATH`
field, the reconciler will look for an existing SAML connection in the
database. If one exists, then the non-unified set of Routes is used for
completing the OAuth flows. If no SAML has been configured, then the
idauth unified Route is used instead.

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>

---------

Signed-off-by: rashmi_kh <rashmi.khanna@in.ibm.com>
Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>
Co-authored-by: Rob Hundley <rwhundle@us.ibm.com>