Reintroduce idauth Route for SAML (#1070)
* Update containers.go to add MASTER_PATH (#1009)

If the `platform-auth-idp` ConfigMap does not have the `MASTER_PATH`
field, the reconciler will look for an existing SAML connection in the
database. If one exists, then the non-unified set of Routes is used for
completing the OAuth flows. If no SAML has been configured, then the
idauth unified Route is used instead.

---------

Signed-off-by: rashmi_kh <rashmi.khanna@in.ibm.com>
Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>
Co-authored-by: Rob Hundley <rwhundle@us.ibm.com>

* Fix testing

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>

* Correct MASTER_PATH field

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>

* Order migration earlier

Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>

---------

Signed-off-by: rashmi_kh <rashmi.khanna@in.ibm.com>
Signed-off-by: Rob Hundley <rwhundle@us.ibm.com>
Co-authored-by: Rashmi Khanna <k.rashmi.r@gmail.com>